# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Achebe net API Documentation",
        "description": "API for Achebe net",
        "contact": {
            "email": "ofuasiasoluchukwu@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://achebe-net-backend.onrender.com/",
            "description": "Achebe net API Server"
        }
    ],
    "paths": {
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user",
                "description": "Get the currently authenticated user.",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "user_property": "user_value"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/connect-wallet": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Route to connect wallet",
                "description": "User register or login by connecting wallet",
                "operationId": "connectWallet",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "wallet_id"
                                ],
                                "properties": {
                                    "wallet_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": "Wallet connected successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "User"
        },
        {
            "name": "Wallet",
            "description": "Wallet"
        }
    ]
}